adding extra security at remove and install

+++ src/remove.rs
@@ -33,16 +33,20 @@ pub fn remove(rawpkg: &String) -> Result<()> {
         let file = fs::read_to_string(format!("/var/lib/pkg/DB/{}/files", rawpkg))?;
         let content = file.lines();
         fs::remove_dir_all(format!("/var/lib/pkg/DB/{}", rawpkg));
-        let protected = if Path::new("/tmp/conflict").exists() {
-            vec!["bin", "lib", "lib64", "sbin", "usr/share/info/dir", "etc/"]
+        let protected = if Path::new("/").exists() {
+            vec!["bin", "lib", "lib64", "sbin", "usr/share/info/dir"]
         } else {
             vec!["bin", "lib", "lib64", "sbin", "usr/share/info/dir"]
         };
         for i in content {
             let to_remove = i.split_whitespace().next().unwrap();
             if !protected.contains(&to_remove) {
-                let _ = fs::remove_file(format!("/{}", to_remove));
-                let _ = fs::remove_dir(format!("/{}", to_remove));
+                if !to_remove.starts_with("etc") {
+                    let _ = fs::remove_file(format!("/{}", to_remove));
+                    let _ = fs::remove_dir(format!("/{}", to_remove));
+                } else {
+                    continue;
+                }
             }
             //println!("Package has been correctly uninstalled !");
         }
